+++ beetsplug/keyfinder.py
@@ -61,9 +61,16 @@ class KeyFinderPlugin(BeetsPlugin):
                 continue
 
             try:
-                output = util.command_output([bin, '-f', item.path])
+                output = util.command_output([bin, b'-f',
+                                              util.syspath(item.path)])
             except (subprocess.CalledProcessError, OSError) as exc:
-                self._log.error(u'execution failed: {0}', exc)
+                self._log.error('execution failed: {0}', exc)
+                continue
+            except UnicodeEncodeError:
+                # Workaround for Python 2 Windows bug.
+                # http://bugs.python.org/issue1759845
+                self._log.error('execution failed for Unicode path: {0!r}',
+                                item.path)
                 continue
 
             key_raw = output.rsplit(None, 1)[-1]